index and event chooser pages swapped and link back to event created on matches page
